+++ frontend/app/src/main/res/layout/activity_waiting_room.xml
@@ -69,7 +69,7 @@
         android:textSize="30sp"
         android:textStyle="bold"
         android:gravity="center"
-
+        app:layout_constraintTop_toTopOf="@id/linear_layout"/>
 
     <LinearLayout
         android:layout_width="match_parent"
@@ -85,7 +85,6 @@
             android:layout_width="match_parent"
             android:layout_height="wrap_content" />
     </LinearLayout>
-
 
     <Button
         android:id="@+id/start_drawing"
